% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{addSeq}
\alias{addSeq}
\title{Add sequence column to GRanges object}
\usage{
addSeq(BSgenome = NULL, intervals)
}
\arguments{
\item{BSgenome}{BSgenome. BSgenome to query sequences, e.g. Hsapiens.}

\item{intervals}{GRanges. GRanges object containing coordinates to query.}
}
\value{
Returns GRanges object containing seqs column
}
\description{
Add sequence column to GRanges object
}
