% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{chunkFiles}
\alias{chunkFiles}
\title{Split files into equally sized chunks}
\usage{
chunkFiles(paths, n_chunks = NULL)
}
\arguments{
\item{paths}{Character. Path(s) to file(s).}

\item{n_chunks}{Integer. Number of groups to split files into.}
}
\description{
Split files into equally sized chunks
}
