% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate.R
\name{curateTEs}
\alias{curateTEs}
\title{Curation of TE intervals}
\usage{
curateTEs(tes, genes = NULL, extend_3p = 0, filt_exonic = FALSE,
  max_e_value = 0.05, min_size = 25)
}
\description{
Curation of TE intervals
}
