% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{deduplicateBAM}
\alias{deduplicateBAM}
\title{Deduplicates SAM/BAM files using umi-tools}
\usage{
deduplicateBAM(bam, samtools = FALSE, suffix = "dedup",
  paired = NULL, outdir = NULL, align_dist = NULL, ncores = 1)
}
\arguments{
\item{bam}{path to single or multiple bam files for combined deduplication}
}
\description{
Deduplicates SAM/BAM files using umi-tools
}
