% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{downsamp}
\alias{downsamp}
\title{Downsamples data in long data.frame format}
\usage{
downsamp(df, i = "id_unique", j = "barcode", n = "n", ds = NULL,
  full = TRUE, seed = 19)
}
\description{
Downsamples data in long data.frame format
}
