% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importBAM}
\alias{importBAM}
\title{Parallel import of BAM files}
\usage{
importBAM(bams, paired = NULL, multi = TRUE, what = NA, tag = NA,
  mate = NA, anchor = NULL, meta = NA, granges = FALSE,
  barcode = NA, data.table = TRUE, sorted = TRUE,
  use_gcluster = FALSE)
}
\arguments{
\item{path}{Character. Path to file.}

\item{main_assembly}{Logical. If TRUE (the default), only returns TE intervals on the main chromosomes.}
}
\value{
Returns a GRanges object with TE intervals
}
\description{
Imports reads stored in BAM files
}
