% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importDFAM}
\alias{importDFAM}
\title{Import TE intervals from DFAM}
\usage{
importDFAM(path, curate = FALSE, main_assembly = TRUE)
}
\arguments{
\item{path}{Character. Path to file.}

\item{curate}{Logical. If TRUE, calls `curateTEs` to resolve overlapping TE intervals.}

\item{main_assembly}{Logical. If TRUE (the default), only returns TE intervals on the main chromosomes.}
}
\value{
Returns a GRanges object with TE intervals
}
\description{
Imports TE interval coordinates from DFAM database
}
