% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importTEs}
\alias{importTEs}
\title{Import TE intervals in common file formats}
\usage{
importTEs(path, curate = FALSE, main_classes = TRUE,
  main_assembly = TRUE, type = NULL, keep_format = TRUE)
}
\description{
Import TE intervals in common file formats
}
