% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importUCSC}
\alias{importUCSC}
\title{Import TE intervals from UCSC table browser}
\usage{
importUCSC(path = "path to UCSC RMSK file", curate = FALSE,
  main_classes = TRUE, main_assembly = TRUE)
}
\arguments{
\item{path}{Character. Path to file.}

\item{curate}{Logical. If TRUE, calls `curateTEs` to resolve overlapping TE intervals.}

\item{main_classes}{Logical. If TRUE (the default), only returns TE intervals unambiguously belonging to DNA, LINE, SINE, and LTR classes.}

\item{main_assembly}{Logical. If TRUE (the default), only returns TE intervals on the main chromosomes.}
}
\value{
Returns a GRanges object with TE intervals
}
\description{
Import TE intervals from UCSC table browser
}
