% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mafft}
\alias{mafft}
\title{Compute multiple sequence alignment}
\usage{
mafft(sequences, ids = NULL, subgroup = FALSE, family = "te",
  save = FALSE, outdir = NULL, threads = -1, epsilon = 0.123,
  memsave = FALSE, fft = TRUE, genafpair = FALSE,
  overwrite = FALSE, op = 1.53, parttree = FALSE, retree = 2,
  maxiterate = 0, long = FALSE, local = FALSE)
}
\arguments{
\item{msa}{DNAStringSet.}
}
\value{
Returns a long data.frame
}
\description{
Computes multiple sequence alignment using mafft.
}
