% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{msaToLong}
\alias{msaToLong}
\title{Transforms multiple sequence alignment into long data.frame}
\usage{
msaToLong(msa)
}
\arguments{
\item{msa}{DNAStringSet.}
}
\value{
Returns a long data.frame
}
\description{
Transforms multiple sequence alignment into long data.frame
}
