% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mutateReads}
\alias{mutateReads}
\title{Randomly mutate read sequences}
\usage{
mutateReads(BSgenome = NULL, reads, tes = NULL, paired = NULL,
  n_reads = NULL, outdir = NULL, prefix = "aligned_reads",
  which = GRanges(), error_rate = NULL, seed = 19)
}
\arguments{
\item{BSgenome}{BSgenome. BSgenome to query sequences, e.g. Hsapiens.}

\item{tes}{GRanges. GRanges object containing TE coordinates}

\item{outdir}{Character. Specifies where to store output fastq files. If NULL (the default), same directory as input bam file.}

\item{bam_path}{Character. Path to input BAM file.}

\item{multi_mapper}{Logical. if FALSE (currently only supported), throws reads mapping to multiple position in the genome (NH:i tag).}
}
\value{
Returns GRanges object containing seqs column
}
\description{
Introduces random mutations in read sequences given a specified error rate. Useful to simulate read mismapping under given error constrains.
}
