% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mutateSeqs}
\alias{mutateSeqs}
\title{Randomly mutate DNA sequences}
\usage{
mutateSeqs(seqs, error_rate = 0, seed = 19)
}
\arguments{
\item{seqs}{DNAStringSet or data.frame with strings.}

\item{error_rate}{Integer. Percentage of base substituations to introduce.}

\item{seed}{Integer. Random seed for sampling.}
}
\value{
Returns DNAStringSet or data.frame with mutated sequences.
}
\description{
Introduces random mutations in DNA sequences given a specified error rate.
}
