% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{writeMM2}
\alias{writeMM2}
\title{Export (sparse) matrix in matrix market format
Much faster than writeMM from Matrix R package}
\usage{
writeMM2(smat, file = NULL)
}
\description{
Export (sparse) matrix in matrix market format
Much faster than writeMM from Matrix R package
}
