% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importBAM}
\alias{importBAM}
\title{Parallel import of BAM files}
\usage{
importBAM(bams, bai_pos = "suffix", paired = NULL, multi = TRUE,
  what = NULL, mate = NA, anchor = NULL, filt = TRUE,
  barcode = "CB")
}
\arguments{
\item{path}{Character. Path to file.}

\item{main_assembly}{Logical. If TRUE (the default), only returns TE intervals on the main chromosomes.}
}
\value{
Returns a GRanges object with TE intervals
}
\description{
Imports reads stored in BAM files
}
